+++ activejob/test/jobs/logging_job.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 class LoggingJob < ActiveJob::Base
-  def perform(dummy)
-    logger.info "Dummy, here is it: #{dummy}"
+  def perform(*args)
+    logger.info "Dummy, here is it: #{args.join(', ')}"
   end
 
   def job_id