+++ activejob/test/jobs/multiple_kwargs_job.rb
@@ -3,7 +3,7 @@
 require_relative "../support/job_buffer"
 
 class MultipleKwargsJob < ActiveJob::Base
-  def perform(argument1:, argument2:)
+  def perform(argument1:, argument2: nil)
     JobBuffer.add("Job with argument1: #{argument1}, argument2: #{argument2}")
   end
 end