+++ activejob/test/jobs/retry_job.rb
@@ -37,7 +37,7 @@ class RetryJob < ActiveJob::Base
   discard_on(CustomDiscardableError) { |job, error| JobBuffer.add("Dealt with a job that was discarded in a custom way. Message: #{error.message}") }
 
   before_enqueue do |job|
-    if job.arguments.include?(:log_scheduled_at) && job.scheduled_at
+    if job.arguments&.include?(:log_scheduled_at) && job.scheduled_at
       JobBuffer.add("Next execution scheduled at #{job.scheduled_at}")
     end
   end