+++ activesupport/lib/active_support/core_ext/method.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require "active_support/core_ext/callable/parameters"
+
+class Method
+  include Callable::Parameters
+end