+++ activejob/CHANGELOG.md
@@ -3,6 +3,11 @@
 
     *Veerpal Brar*
 
+*   Raise an `ArgumentErrorError` if `perform_later` is invoked with arguments
+    not compatible with `perform`. 
+
+    *Christian Schmidt*
+
 
 ## Rails 7.0.0.alpha2 (September 15, 2021) ##